Fixes service issues - no more crashes after start of a service with battery optimization enabled - no more service restarts after the app exit with disabled private notifications
